fix: fixing (almost) all warnings triggered when compiling tests & demos (#163)
- [x] I have installed [pre-commit](https://pre-commit.com/) locally and
use it to validate my commits.
- [x] The PR title follows the [conventional
commits](https://www.conventionalcommits.org/en/v1.0.0/) convention.
Available tags: 'build', 'chore', 'ci', 'docs', 'feat', 'fix', 'perf',
'refactor', 'revert', 'style', 'test'
- [x] This new PR is tested.

## Description
Trying to fix as many warnings as possible with little change to the
code: renaming variable, fixing type alias or adding new one, remove
useless cast, ...

Tested with GCC 11.4, 12.3 and LLVM 14.0.

## Related issue
None

## How has this been tested?
Launching the already existing tests.

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [x] I agree to follow this project's Code of Conduct

+++ demos/multigrid/main.cpp
@@ -103,7 +103,7 @@ template <class Mesh>
     min_level = level - 1;
     max_level = level;
 
-    int i = static_cast<int>(1 << min_level);
+    int i = 1 << min_level;
 
     cl_type cl;
     if constexpr (Mesh::dim == 1)
